do HQ API call synchronously
